refactor(platform): extract shared provider form fields component
Reduce code duplication between OnboardingModal and ProviderSetupPage by
extracting provider type selector and form fields into a shared
ProviderFormFields component.
